fix(capa1): SPEC-capa1 review-C — ORM model corrections
H-2 (pr-review-toolkit:code-reviewer): Transcription.duration_seconds
ahora es Mapped[Decimal] en vez de Mapped[float]. Numeric(10,2) +
asyncpg + SQLAlchemy materializan como decimal.Decimal en runtime; la
anotacion anterior mentia a mypy y rompia aritmetica mixta float/Decimal
con TypeError. Test test_transcription_duration_round_trips_as_decimal
verifica el round-trip explicitamente.

H-7 (sandinas-code-reviewer): OAuthToken.updated_at ahora tiene
onupdate=func.clock_timestamp(). server_default=func.now() solo dispara
en INSERT, congelando el campo desde la primera escritura. Capa 2
rotara tokens y depende de updated_at avanzando.

Detalle de implementacion: clock_timestamp() en lugar de now(). Postgres
now() = transaction_timestamp() — dos UPDATEs en la misma transaccion
comparten timestamp. clock_timestamp() devuelve wall-time por call,
asegurando que cada UPDATE tenga su tiempo real. Solo se usa en
onupdate; server_default mantiene now() (el valor en INSERT representa
"esta transaccion" lo cual es semantica correcta para createdAt).

Tests adicionales (pr-test-analyzer #3 — JSONB nested round-trip):
- test_transcription_jsonb_round_trips_nested: valida que estructuras
  JSONB con arrays anidados, floats, unicode (incluyendo "¿cómo estás?"),
  speakers y words round-trip losslessly. La factory anterior solo
  exercitaba {"segments": []} — caso degenerado.

Tests: 85/85 verde (+3).

+++ src/transcription_api/db/models/transcription.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import Any
 
 from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, Numeric, Text, Uuid, func, text
@@ -34,7 +35,9 @@ class Transcription(Base):
     audio_hash: Mapped[str] = mapped_column(Text, nullable=False, index=True)
     original_filename: Mapped[str] = mapped_column(Text, nullable=False)
     original_size_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    duration_seconds: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    # H-2: Numeric(10,2) returns Decimal at runtime (asyncpg + SQLAlchemy);
+    # using Mapped[float] would lie to mypy and break Decimal-aware arithmetic.
+    duration_seconds: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     language: Mapped[str] = mapped_column(Text, nullable=False)
     num_speakers: Mapped[int] = mapped_column(Integer, nullable=False)
     text_content: Mapped[str] = mapped_column("text", Text, nullable=False)
